+++ src/components/Profile/ProfilePage/ProfilePageSections/ProfilePageSections.tsx
@@ -8,7 +8,7 @@ import GestureRecognizer, {
 import { ProfileHeaderContainer } from "./ProfileHeaderContainer";
 import { ProfileDetailSection } from "./ProfileDetailSection";
 import { white } from "../../../../themeColors";
-import { VerifyProfile } from "../../VerifyProfile";
+import { VerifyProfile } from "../../../Preview/VerifyProfile";
 import { VerifyModal } from "../../../Modals/VerifyModal";
 import { profileTypeEnum } from "../../profileTypeEnum";
